Rename concrete implementation to TempFolderInfo
